Added ESC disambiguation (literal ESC vs. a terminal sequence)
This change allows jline to tell the difference between a
literal ESC (the user hitting the ESC key) and an ESC that
is part of, say, a terminal control sequence.  This is done
in the same way that vi does it: when an ESC is encountered
we wait for a short period of time (150ms -- controllable by
jline.esc.timeout) to see if another character is coming. If
one isn't, then it is a literal ESC, otherwise we follow jline's
normal behavior.

The one note-worthy introduction of this code is a new thread:
Standard java I/O doesn't have the ability to peek into a stream
in a non-blocking fashion, so this peeking is performed by
a separate thread which will block indefinitely until the byte
is read. The thread is only started if jline.esc.timeout is
a positive value, so it can be disabled with this setting. The
thread is only active immediately after an ESC is received
in which case it is responsible for reading the next keystroke.
After that keystroke is read, all other reads happen in the
context of the main jline thread.

There is a new shutdown() method on ConsoleReader() that shuts
the reader down and, more specifically, shuts the helper thread
down.  Also the ConsoleReader() has a finalizer that should
shut the thread down if it gets cleaned up.