ConsoleKeys is now the one-true-source of keymaps.
While it was possible to set the current keymap, it
was almost impossible to find out what the current keymap was
and there were all sorts of places where the tracking of whether
vi mode was active or inactive was broken.  This change centralizes
all of this in the ConsoleKeys class.  KeyMaps can now be set
and retrieve by name and a map knows its name and knows whether or
not it considered a "vi" keymap, making it almost impossible to get
wrong.

I also exposed API's in ConsoleReader() so that the client can
switch between keymaps (this was necessary for better jsqsh
integration).