don't print an error when failing to query stty information
When trying to execute stty to get terminal properties the code can trigger
an InterruptedException even though we've read all of the output from stty
or at least enough for what we need. Instead of printing a warning and
returning -1 in this case instead try to parse what data we do have and
reset the cache timer.