Fix switch statement in test class
Add a missing default case for switch statement, which throws the illegal
argument exception properly. This is more explicit than falling through an
undeclared empty default case, and then throwing the exception outside the
switch statement. This cleans up a warning in some IDEs which check switch
statements for problems.