When adding spaces in place of tabs, calculate from start of line
Maintain the length of the last line in AttributedStringBuilder so that
future appends also calculate tab size correctly.

If there are line breaks in a string, a tab character should add spaces
up until the next tabstop from the beginning of the line not from the
start of the string.