Avoid encoding console input on Windows when using Terminal.reader()
On Windows, console input is read using the Windows API as UTF-16
characters. When using the provided Terminal.reader(), input is
encoded using UTF-8 just to decode it again shortly after.

Similar to the changes made in #168, change Terminal.reader() to
consume the characters directly to avoid unnecessary encoding.

Instead of a PipedInput/OutputStream a new PumpReader is now used
that works very much like a PipedReader/Writer. However, it
additionally supports reading input using an InputStream with
a specific Charset. It's much easier and more reliable to
implement this directly in the Reader because you need to buffer
to encode characters correctly.