Windows: Use virtual terminal processing if available
Windows 10 added native support for ANSI escape codes. However, that
functionality is hidden behind the ENABLE_VIRTUAL_TERMINAL_PROCESSING
flag that needs to be set manually using the Windows API.

Use the native functionality on Windows 10, but fall back to the
AnsiWriter if setting the flag fails on older Windows versions.