Clear internal history in case of currupted histroy file
The expected behaviour when loading of a history files fails is that the issue is logged but the reader is usable.

The current behaviour results in a IllegalArgumentException (from the addHistoryLine(Path, String) function) which is never caught. This results is a different behaviour than for example the file permission does not allow to read the file.
As a consequence in Karaf the session is not useable because of returned "null" at some point which is interpreted as session end.

This commit introduces a fix which catches the IllegalArgumentException and keeps the reader usable without history information. Also this makes the behaviour consistent.