Add support for embedded applications
There are unusual situations in which a container application may be
using JLine3 and be invoking an application that also uses JLine3. When
this is the case, there can be conflicts between the container's JLine
usage and the application. To be more specific, this is a problem for
sbt and the scala REPL. This is a difficult problem to solve because sbt
attempts to support all versions of scala but there is no entry point to
the scala REPL that allows the calling application to pass in a custom
terminal. As a result, it isn't possible for sbt using JLine3 to viably
support all versions of the scala REPL.

One possible workaround to solve the sbt use case is to allows the application
to set a global Terminal in TerminalBuilder that will be returned by all
TerminalBuilder instances if it is set.