chore: Add DevContainer JSON
This makes the contributor experience a lot smoother, because it enables e.g. GitHub Codespaces to "just work" - with the correct more recent than default Java version already pre-installed, and the build "warmed up" (thanks to the postCreateCommand).

This is technically not specific only to GitHub Codespaces, but also useful to have a clean standard development environment in e.g. for both Visual Studio Code as well as IntelliJ, both of which nowadays also support https://containers.dev.