fixes line reader setting options via system properties
Attempted to set the line reader option DISABLE_EVENT_EXPANSION by
setting the java system property
`-Dorg.jline.reader.props.disable-event-expansion=on` and it did not
work.  The code for processing this option is comparing the enum name
to the system property value when it should be looking at the key.